ai-generator/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.43.0
werkzeug==3.0.3